moved customthemesprovider to a folder so we can break things into components; fixed imports (and moved an import to the right spot in index.js)
